// src/modules/bloggers-platform/application/dto/comments/comment-response.dto.ts
import { LikeStatus } from '../../../domain/likes/like.dto';

export interface ICommentResponseDto {
  id: string;
  content: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  createdAt: string;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
}